usb-pad/evdev: de-duplicated axis correction code
usb-pad/joydev: dedup device_data
