+++ src/usb-pad/evdev/shared.h
@@ -1,4 +1,6 @@
 #pragma once
+#include <linux/joystick.h>
+#include <unistd.h>
 #include "gtk.h"
 #include "../padproxy.h"
 #include "../../configuration.h"
@@ -102,6 +104,29 @@ struct ConfigData
 	ApiCallbacks *cb;
 	bool use_hidraw_ff_pt;
 	const char *dev_type;
+};
+
+struct axis_correct
+{
+	int used;
+	int coef[3];
+};
+
+struct device_data
+{
+	int fd;
+	std::string name;
+	uint8_t axis_map[ABS_MAX + 1];
+	bool axis_inverted[3];
+	int axis_initial[3];
+	uint16_t btn_map[KEY_MAX + 1];
+	struct axis_correct abs_correct[ABS_MAX];
+	int axes = 0;
+	int buttons = 0;
+	std::vector<uint16_t> mappings;
+	bool is_gamepad; //xboxish gamepad
+	bool is_dualanalog; // tricky, have to read the AXIS_RZ somehow and
+					// determine if its unpressed value is zero
 };
 
 int GtkPadConfigure(int port, const char* dev_type, const char *title, const char *apiname, GtkWindow *parent, ApiCallbacks& apicbs);
